split tour code
